+++ inspeximus/__init__.py
@@ -66,6 +66,7 @@ from .actions import ActionLedger
 from .subject_rights import export_subject, rectify
 from .technical_documentation import annex_iv, instructions_for_use, registration_export
 from .deployer import deployer_report, dpia_appendix, fria_appendix
+from .agent_audit_trail import export_jsonl as export_audit_trail, verify_jsonl as verify_audit_trail
 
 __all__ = [
     "Inspeximus",
@@ -78,6 +79,8 @@ __all__ = [
     "deployer_report",
     "dpia_appendix",
     "fria_appendix",
+    "export_audit_trail",
+    "verify_audit_trail",
     "AmbiguousSubject",
     "new_receipt_keypair",
     "receipt_key_for",